api/requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
numpy
python-multipart
